Small refactor to make code usable
