+++ Cosmetics/Cosmetics/src/main/resources/plugin.yml
@@ -1,0 +1,7 @@
+name: Emotes
+version: 1.0
+author: ThinkPuppy
+main: me.thinkpuppy.emotes.Main
+description: It's a emotes plugin which lets you express your current emotions.
+
+commands:
